+++ Configuration/TSconfig/GridElements.ts
@@ -3,5 +3,6 @@
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/AccordionGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ColumnLayoutGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/DocumentSectionGridElement.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/PanelGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/TimelineGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ViewportGridElement.ts">