+++ Configuration/TypoScript/GridElements/constants.ts
@@ -3,5 +3,6 @@
 <INCLUDE_TYPOSCRIPT: source="FILE: ./AccordionGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./DocumentSectionGridElement/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./PanelGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./TimelineGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ViewportGridElement/constants.ts">