+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `Panel` Grid Element
 * Added `Timeline` and `Event` Grid Element
 * Updated CSS class of Grid Element Headers
 * Updated `Table` layouts and partials